requirements.txt
# CWtoSDP - ConnectWise to ServiceDesk Plus Integration
# Python dependencies

requests>=2.28.0
pandas>=2.0.0
python-dotenv>=1.0.0

